+++ files/security/Credentials.groovy
@@ -4,10 +4,13 @@ import com.cloudbees.plugins.credentials.domains.Domain
 import com.cloudbees.plugins.credentials.impl.*
 import com.cloudbees.plugins.credentials.CredentialsScope
 import com.cloudbees.plugins.credentials.SystemCredentialsProvider
+import com.cloudbees.plugins.credentials.SecretBytes
 import org.jenkinsci.plugins.plaincredentials.impl.StringCredentialsImpl
+import org.jenkinsci.plugins.plaincredentials.impl.FileCredentialsImpl
 import hudson.util.Secret
 import hudson.security.HudsonPrivateSecurityRealm
 import jenkins.model.Jenkins
+import java.nio.file.*
 
 def ansible = new Domain("ansible", "ansible managed credentials", null)
 SystemCredentialsProvider.getInstance().getStore().addDomain(ansible)
@@ -27,6 +30,14 @@ if ("${cred}" != "") {
         "${id}",
         "${description}",
         Secret.fromString("${text}"))
+} else if ("${file}" != "") {
+    Path fileLocation = Paths.get("${file}");
+    credential = new FileCredentialsImpl(
+        CredentialsScope.GLOBAL, 
+        "${id}", 
+        "${description}", 
+        fileLocation.getFileName().toString(), 
+        SecretBytes.fromBytes(Files.readAllBytes(fileLocation)))
 } else {
     credential = new UsernamePasswordCredentialsImpl(
         CredentialsScope.GLOBAL,