+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a changelog](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased](https://github.com/idealista/jenkins_role/tree/develop)
+### Added
+- *[#69](https://github.com/idealista/jenkins_role/issues/69) [FEATURE] Allow to store secret file credentials* @emepege
+
 ## [2.7.0](https://github.com/idealista/jenkins_role/tree/2.7.0)
 ## [Full Changelog](https://github.com/idealista/jenkins_role/compare/2.6.0...2.7.0)
 ### Added